+++ pkgs/ffigen/CHANGELOG.md
@@ -6,6 +6,8 @@
   - https://github.com/dart-lang/native/issues/1582
   - https://github.com/dart-lang/native/issues/1594
   - https://github.com/dart-lang/native/issues/1595
+- Allow static and instance methods to have the same name:
+  https://github.com/dart-lang/native/issues/1136
 - __Breaking change__: Change the way ObjC categories are generated. Instead of
   inserting their methods into the interface, generate Dart extension methods.
   For instance methods this makes no difference to user code (as long as the